Finish up provider stuff
