+++ Essentials/pom.xml
@@ -68,32 +68,26 @@
         </dependency>
         <dependency>
             <groupId>net.ess3</groupId>
-            <artifactId>NMSProvider</artifactId>
+            <artifactId>BaseProviders</artifactId>
             <version>2.17.2</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>net.ess3</groupId>
-            <artifactId>UpdatedMetaProvider</artifactId>
+            <artifactId>NMSReflectionProvider</artifactId>
             <version>2.17.2</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>net.ess3</groupId>
-            <artifactId>LegacyProvider</artifactId>
+            <artifactId>PaperProvider</artifactId>
             <version>2.17.2</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>net.ess3</groupId>
-            <artifactId>ReflectionProvider</artifactId>
-            <version>2.17.2</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>net.ess3</groupId>
-            <artifactId>FlattenedProvider</artifactId>
-            <version>2.17.2</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>com.destroystokyo.paper</groupId>
+                    <artifactId>paper-api</artifactId>
+                </exclusion>
+            </exclusions>
             <scope>compile</scope>
         </dependency>
     </dependencies>
