+++ Essentials/src/com/earth2me/essentials/IEssentials.java
@@ -6,7 +6,8 @@ import com.earth2me.essentials.api.IWarps;
 import com.earth2me.essentials.metrics.Metrics;
 import com.earth2me.essentials.perm.PermissionsHandler;
 import com.earth2me.essentials.register.payment.Methods;
-import net.ess3.nms.SpawnerProvider;
+import net.ess3.provider.ServerStateProvider;
+import net.ess3.provider.SpawnerProvider;
 import org.bukkit.World;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -106,4 +107,6 @@ public interface IEssentials extends Plugin {
     Iterable<User> getOnlineUsers();
 
     SpawnerProvider getSpawnerProvider();
+
+    ServerStateProvider getServerStateProvider();
 }
