+++ Essentials/src/net/ess3/api/IEssentials.java
@@ -1,8 +1,8 @@
 package net.ess3.api;
 
 import com.earth2me.essentials.items.CustomItemResolver;
-import net.ess3.nms.PotionMetaProvider;
-import net.ess3.nms.SpawnEggProvider;
+import net.ess3.provider.PotionMetaProvider;
+import net.ess3.provider.SpawnEggProvider;
 
 import java.util.Collection;
