--- nms/NMSProvider/src/net/ess3/providers/Provider.java
@@ -1,7 +1,0 @@
-package net.ess3.providers;
-
-public interface Provider {
-    boolean tryProvider();
-
-    String getDescription();
-}
