+++ providers/BaseProviders/src/net/ess3/provider/PotionMetaProvider.java
@@ -1,0 +1,18 @@
+package net.ess3.provider;
+
+import org.bukkit.Material;
+import org.bukkit.inventory.ItemStack;
+
+public interface PotionMetaProvider extends Provider {
+    ItemStack createPotionItem(Material initial, int effectId);
+
+    @Override
+    default boolean tryProvider() {
+        try {
+            createPotionItem(Material.POTION, 8260); // Poison Level II Extended
+            return true;
+        } catch (Throwable t) {
+            return false;
+        }
+    }
+}
