+++ providers/BaseProviders/src/net/ess3/provider/ServerStateProvider.java
@@ -1,0 +1,10 @@
+package net.ess3.provider;
+
+public interface ServerStateProvider extends Provider {
+    boolean isStopping();
+
+    @Override
+    default boolean tryProvider() {
+        return false;
+    }
+}
