+++ providers/BaseProviders/src/net/ess3/provider/providers/LegacyPotionMetaProvider.java
@@ -1,0 +1,20 @@
+package net.ess3.provider.providers;
+
+import net.ess3.provider.PotionMetaProvider;
+import org.bukkit.Material;
+import org.bukkit.inventory.ItemStack;
+
+@SuppressWarnings("deprecation")
+public class LegacyPotionMetaProvider implements PotionMetaProvider {
+    @Override
+    public ItemStack createPotionItem(Material initial, int effectId) {
+        ItemStack potion = new ItemStack(initial, 1);
+        potion.setDurability((short) effectId);
+        return potion;
+    }
+
+    @Override
+    public String getDescription() {
+        return "Legacy 1.8 Potion Meta Provider";
+    }
+}
