--- providers/BaseProviders/src/net/ess3/provider/providers/SpawnerProvider.java
@@ -1,6 +1,0 @@
-package net.ess3.provider.providers;
-
-import net.ess3.provider.Provider;
-
-public interface SpawnerProvider extends Provider {
-}
